inventory-service/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
sqlalchemy==2.0.21
kafka-python==2.0.2
python-dotenv==1.0.0